setting SparkManager as a instance variable in SustainServer instead of a static variable - cleaner approach
